refactor(read): extract shared display components for reuse
Move formatTime, getFontSize, segmentClass, ORPDisplay, CenteredDisplay, ContextDisplay, and SourcePanel into a shared components.tsx so the upcoming book reader at /read/book can import them directly.
